requirements.txt
airtable-python-wrapper==0.12.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.2.0
azure-common==1.1.23
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.9.221
botocore==1.12.221
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
docutils==0.15.2
future==0.17.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
googleapis-common-protos==1.3.5
httplib2==0.13.1
idna==2.8
ijson==2.4
importlib-metadata==0.23
jmespath==0.9.4
lookerapi==3.0.0
more-itertools==7.2.0
numpy==1.17.1
oauth2client==4.1.3
oauthlib==3.1.0
packaging==19.2
pandas==0.25.1
pluggy==0.13.0
prefect==0.7.0
protobuf==3.9.1
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pycryptodomex==3.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
pytest==5.2.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
snowflake==0.0.3
snowflake-connector-python==1.8.7
snowflake-sqlalchemy==1.1.14
SQLAlchemy==1.3.8
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.6.0
